// modules/review/client-react/graphql/__generated__/deleteReview.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: deleteReview
// ====================================================

export interface deleteReview {
  deleteReview: boolean | null;
}

export interface deleteReviewVariables {
  id: number;
}