// modules/review/client-react/graphql/__generated__/addReview.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { AddModalReviewInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: addReview
// ====================================================

export interface addReview {
  addReview: boolean | null;
}

export interface addReviewVariables {
  input: AddModalReviewInput;
}