// modules/review/client-react/graphql/__generated__/editReview.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { EditReviewInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: editReview
// ====================================================

export interface editReview {
  editReview: boolean | null;
}

export interface editReviewVariables {
  input: EditReviewInput;
}