// modules/review/client-react/graphql/__generated__/reviewHelpfulStatus.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: reviewHelpfulStatus
// ====================================================

export interface reviewHelpfulStatus {
  reviewHelpfulStatus: boolean | null;
}

export interface reviewHelpfulStatusVariables {
  reviewId?: number | null;
  userId?: number | null;
}