// modules/review/client-react/graphql/__generated__/addOrRemoveReviewHelpful.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: addOrRemoveReviewHelpful
// ====================================================

export interface addOrRemoveReviewHelpful {
  addOrRemoveReviewHelpful: string | null;
}

export interface addOrRemoveReviewHelpfulVariables {
  reviewId?: number | null;
  userId?: number | null;
}